Lauren Acosta, CEN 3024C
This Java program is a simple random number generator game with a menu-driven interface. The user is prompted to add the values in the generated array. The program uses a variety of methods (getRandomNumber, printArray, playGame, shuffle, newValues, actualSum, compare, swap, generatePos) to organize and implement different functionalities. Overall, the program provides a simple interactive experience where the user can play a game, shuffle values, generate new values, sort the array, and exit the program. Example of input and output : 
Select your option:
1. Play Game
2. Swap Values
3. New Values
4. Sort
5. Exit

User Input: 1

Add these values:
10 17 8 15
User Input: 50

Output:
That was incorrect. The value adds to be: 50